Fixed stack-smashing bug in the signing/verification tests (it was due to an attempt to write a size_t value to a location that held an unsigned int).

+++ lib/test/test_util.c
@@ -546,7 +546,7 @@ END_TEST
 START_TEST(test_sign_openssl_small)
 {
     unsigned char *signature;
-    unsigned int signatureLen = 0;
+    size_t signatureLen = 0;
     unsigned int size;
     int ret;
